Derive Macro snarky_deriver::SnarkyType

#[derive(SnarkyType)]
{
    // Attributes available to this derive:
    #[snarky]
}

The SnarkyType derive macro. It generates implementations of [kimchi::snarky::SnarkyType], as long as your structure’s fields implement that type as well. It works very similarly to [serde].

For example:

#[derive(kimchi::SnarkyType)]
struct MyType<F> where F: PrimeField {
  // ...
}

You can specify your snarky type implementation to only apply to a single field (e.g. Fp), or to apply to a field that has a different name than F:

#[derive(kimchi::SnarkyType)]
#[snarky(field = "G::ScalarField")]
struct MyType<G> where G: KimchiCurve {

You can skip a field in the serializer:

#[derive(kimchi::SnarkyType)]
struct MyType<F> where F: PrimeField {
 #[snarky(skip)]
field_to_skip: NotASnarkyType,

Finally, you can specify a custom check function, as well as a custom auxiliary function and type:

#[derive(kimchi::SnarkyType)]
#[snarky(check_fn = "my_check_fn")]
#[snarky(auxiliary_fn = "my_auxiliary_fn")]
#[snarky(auxiliary_type = "MyAuxiliaryType")]
struct MyType<F> where F: PrimeField {

By default, a tuple will be use to represent the out-of-circuit type. You can specify it yourself by using the value helper attribute:

#[derive(kimchi::SnarkyType)]
#[snarky(value = "MyOutOfCircuitType")]
struct MyType<F> where F: PrimeField {

and implement the [CircuitAndValue] trait on your [SnarkyType].